gosec fixes; int64 proto fields;

+++ go/integrations/integration_promote_user_test.go
@@ -60,7 +60,7 @@ func testIntegrationPromoteUser(t *testing.T) {
 			t.Errorf("failed to get new token after user patches self %v", err)
 		}
 
-		if session.RoleBits&int32(types.SiteRoles_APP_GROUP_ADMIN) > 0 {
+		if session.RoleBits&int64(types.SiteRoles_APP_GROUP_ADMIN) > 0 {
 			t.Error("user has admin role after trying to add admin role to themselves")
 		}
 	})
@@ -77,7 +77,7 @@ func testIntegrationPromoteUser(t *testing.T) {
 			t.Errorf("failed to get new token after admin promotes staff %v", err)
 		}
 
-		if session.RoleBits&int32(types.SiteRoles_APP_GROUP_SCHEDULES) == 0 {
+		if session.RoleBits&int64(types.SiteRoles_APP_GROUP_SCHEDULES) == 0 {
 			t.Error("staff does not have APP_GROUP_SCHEDULES after admin promotion")
 		}
 	})
@@ -93,7 +93,7 @@ func testIntegrationPromoteUser(t *testing.T) {
 			t.Errorf("failed to get new token after staff modify staff role without permissions %v", err)
 		}
 
-		if session.RoleBits&int32(types.SiteRoles_APP_GROUP_SCHEDULES) > 0 {
+		if session.RoleBits&int64(types.SiteRoles_APP_GROUP_SCHEDULES) > 0 {
 			t.Error("staff modified staff role without having APP_GROUP_USERS permissions")
 		}
 
@@ -107,7 +107,7 @@ func testIntegrationPromoteUser(t *testing.T) {
 			t.Errorf("failed to get new token after staff modify user role without permissions %v", err)
 		}
 
-		if session.RoleBits&int32(types.SiteRoles_APP_GROUP_USERS) == 0 {
+		if session.RoleBits&int64(types.SiteRoles_APP_GROUP_USERS) == 0 {
 			t.Error("staff does not have APP_GROUP_USERS permissions after admin add")
 		}
 
@@ -128,7 +128,7 @@ func testIntegrationPromoteUser(t *testing.T) {
 			t.Errorf("failed to get new token after staff modify user role with permissions %v", err)
 		}
 
-		if session.RoleBits&int32(types.SiteRoles_APP_GROUP_SCHEDULES) == 0 {
+		if session.RoleBits&int64(types.SiteRoles_APP_GROUP_SCHEDULES) == 0 {
 			t.Error("staff failed modify user role having APP_GROUP_USERS permissions")
 		}
 
@@ -146,10 +146,10 @@ func testIntegrationPromoteUser(t *testing.T) {
 			if err != nil {
 				t.Errorf("failed to get new staff token after role update staffid:%s %v", staff.TestUserId, err)
 			}
-			if session.RoleBits&int32(types.SiteRoles_APP_GROUP_SCHEDULES) == 0 {
+			if session.RoleBits&int64(types.SiteRoles_APP_GROUP_SCHEDULES) == 0 {
 				t.Errorf("staff %s does not have APP_GROUP_SCHEDULES permissions, %d", staff.TestUserId, session.RoleBits)
 			}
-			if session.RoleBits&int32(types.SiteRoles_APP_GROUP_USERS) == 0 {
+			if session.RoleBits&int64(types.SiteRoles_APP_GROUP_USERS) == 0 {
 				t.Errorf("staff %s does not have APP_GROUP_USERS permissions, %d", staff.TestUserId, session.RoleBits)
 			}
 			testUserIdx := int32(i + 1) // skip admin
@@ -166,7 +166,7 @@ func testIntegrationPromoteUser(t *testing.T) {
 			if err != nil {
 				t.Errorf("failed to get new member token after role update memberid:%s %v", member.TestUserId, err)
 			}
-			if session.RoleBits&int32(types.SiteRoles_APP_GROUP_BOOKINGS) == 0 {
+			if session.RoleBits&int64(types.SiteRoles_APP_GROUP_BOOKINGS) == 0 {
 				t.Errorf("member %s does not have APP_GROUP_BOOKINGS permissions, %d", member.TestUserId, session.RoleBits)
 			}
 			testUserIdx := int32(i + 4) // skip admin + 3 staff
